synapse-nexus-web/backend/requirements.txt
fastapi
uvicorn[standard]
pyjwt
python-dotenv
pydantic[email]
